types fix

+++ src/parser.c
@@ -7,9 +7,9 @@
 
 /* CDT del parser */
 struct parser {
-    /** tipificación para cada caracter */
+    /** tipificaciÃ³n para cada caracter */
     const unsigned     *classes;
-    /** definición de estados */
+    /** definiciÃ³n de estados */
     const struct parser_definition *def;
 
     /* estado actual */
@@ -60,7 +60,7 @@ parser_feed(struct parser *p, const uint8_t c) {
         const int when = state[i].when;
         if (state[i].when <= 0xFF) {
             matched = (c == when);
-        } else if(state[i].when == ANY) {
+        } else if(state[i].when == (int) ANY) {
             matched = true;
         } else if(state[i].when > 0xFF) {
             matched = (type & when);
@@ -88,4 +88,3 @@ const unsigned *
 parser_no_classes(void) {
     return classes;
 }
-
